+++ src/commands/index.ts
@@ -6,9 +6,12 @@ import openWebview from './open-webview'
 import resetSettings from './reset-settings'
 import selectLanguage from './select-language'
 import showChannel from './show-channel'
+import svgo from './svgo'
 
 const m: ExtensionModule = (ctx) => {
-  return flatten([openWebview, resetSettings, showChannel, mirror, selectLanguage, clearCache].map((fn) => fn(ctx)))
+  return flatten(
+    [openWebview, resetSettings, showChannel, mirror, selectLanguage, clearCache, svgo].map((fn) => fn(ctx)),
+  )
 }
 
 export * from './commands'