+++ src/core/operator/index.ts
@@ -1,4 +1,4 @@
 export * from './compressor'
 export * from './format-converter'
 export * from './operator'
-export * from './svgo'
+export * from './svgo/svgo'